requirements.txt
streamlit
python-docx
pdfminer.six
scikit-learn
fpdf